warehouse_dms/requirements.txt
Django==5.1.4
django-ninja==1.4.3
psycopg2-binary==2.9.10
pgvector==0.3.6
celery==5.4.0
django-celery-beat==2.8.1
redis==5.2.1
python-dotenv==1.0.1
pycryptodome==3.21.0
PyJWT==2.10.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-storages[google]==1.14.4
dj-database-url==2.3.0
gunicorn==23.0.0
uvicorn==0.32.1
whitenoise==6.8.2
scalar-django-ninja==0.1.4
google-auth==2.37.0
africastalking==1.2.7
Jinja2==3.1.4
google-cloud-vision==3.1.0
google-cloud-aiplatform==1.71.1
